pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "gov-uk-mcp"
version = "0.1.0"
description = "MCP server for querying UK government data and services"
readme = "README.md"
requires-python = ">=3.10"
license = "MIT"
authors = [
    {name = "Steve Chan"}
]
dependencies = [
    "mcp>=1.0.0",
    "requests>=2.31.0",
    "python-dotenv>=1.0.0"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "pre-commit>=3.5.0",
    "detect-secrets>=1.4.0"
]

[project.scripts]
gov-uk-mcp = "gov_uk_mcp.server:main"

[tool.setuptools]
packages = ["gov_uk_mcp", "gov_uk_mcp.tools"]

[tool.setuptools.package-data]
gov_uk_mcp = ["py.typed"]

[tool.black]
line-length = 100
target-version = ['py310']

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
